整合子服务层逻辑到主服务层
